Factor LhsTensorSymmAndIndices into separate file
